styling(fonts): change font-display to 'fallback'
In order to get a better UX on slow mobile connections, we change the
Averta @font-face to use 'fallback' instead of the default which is most
likely 'block'. Additionally we set 'sans-serif' as a fallback font when
Averta loads too slow.

See also https://developers.google.com/web/updates/2016/02/font-display

Closes https://app.asana.com/0/995964219886332/1118204062378468